refactor: divide follow-profile component into two
